backend/requirements.txt
blinker==1.9.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
click==8.3.1
cryptography==46.0.6
dotenv==0.9.9
Flask==3.1.0
flask-cors==6.0.2
Flask-JWT-Extended==4.7.1
gunicorn==25.3.0
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
packaging==26.0
pycparser==3.0
PyJWT==2.12.1
PyMySQL==1.1.2
python-dotenv==1.2.1
requests==2.31.0
urllib3==2.6.3
Werkzeug==3.1.6
